+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.31.0...master)
+## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.32.0...master)
+
+
+## [0.32.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.32.0...master)
 ### Added
 - Daily value graph now handles masked data
 - The timeSeriesId is now a parameter on the graph image server to show a specific time series for